% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluarPosicion.R
\docType{data}
\name{bitboards}
\alias{bitboards}
\title{evaluación estática de una posición representada en tablero}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 7 columns.
}
\usage{
bitboards
}
\arguments{
\item{tablero}{matriz que representa el estado de un tablero}

\item{turno}{qué jugador comienza pa partida: 1 para humano, 2 para IA.}
}
\description{
evalua una posición mediante criterios estáticos basados en el
número de casillas conectadas del jugador en turno
}
\examples{
tablero <- crear_posicion_aleatoria(18)
visualizar_tablero(tablero)
evaluar_posicion(tablero)
.evaluar_turno(tablero, 1)
.evaluar_turno(tablero, 2)
}
\keyword{datasets}
