% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profundidadAdaptativa.R
\name{.columnasValidas}
\alias{.columnasValidas}
\title{profundidad adaptativa}
\usage{
.columnasValidas(tablero)
}
\description{
profundidad adaptativa
}
\examples{
tablero <- crear_posicion_aleatoria(10)
.adaptativa(tablero)
}
