% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generarPosicion.R
\name{.crear_posicion_aleatoria}
\alias{.crear_posicion_aleatoria}
\title{Crear posición aleatoria}
\usage{
.crear_posicion_aleatoria(profundidad = 10, turno = 1)
}
\arguments{
\item{tablero}{matriz 6 x 7 que representa la situación del tablero de juego.}
}
\description{
dada una situacvión en el tablero de juego, devuelve las posibles
  jugadas existentes: columnas no completadas
}
\examples{
tablero <- .crear_posicion_aleatoria(42, 1)
visualizar_tablero(tablero)
}
