% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jugadas_disponibles_r}
\alias{jugadas_disponibles_r}
\title{Columnas disponibles (C++)}
\usage{
jugadas_disponibles_r(tablero)
}
\arguments{
\item{tablero}{IntegerMatrix 6×7. Estado del tablero.}
}
\value{
IntegerVector con los índices de columna disponibles (1-indexados).
}
\description{
Versión C++ de \code{\link{jugadas_disponibles}}. Devuelve los
  índices (1-7) de las columnas que no están completamente llenas.
}
\examples{
tablero <- crear_posicion_aleatoria(20)
jugadas_disponibles_r(tablero)
jugadas_disponibles(tablero)   # versión R equivalente

}
\seealso{
\code{\link{jugadas_disponibles}}, \code{\link{ordenar_jugadas_r}}
}
