% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnoHumano.R
\name{turno_humano}
\alias{turno_humano}
\title{solicita una jugada al humano.}
\usage{
turno_humano(tablero, jugada = NULL)
}
\arguments{
\item{tablero}{matriz 6 x 7 que representa la situación del tablero de juego.}

\item{jugada}{un entero del 1 al 7 correspondiente a cada una de las 7 columnas
si jugada = NULL (valor por defecto), entonces invita al jugador a introducir su jugada}
}
\description{
la función invita al jugador humano a introducir una jugada
Si la jugada es legal representeala jugada del humano en el tablero mediante
un 1 en la casilla que corresponda. La jugada de la IA será representada en el
tablero mediante un 2.
}
\details{
# Si la jugada es legal representa la jugada del humano en el tablero mediante un 1
en la casilla que corresponda a la jugada
Si la jugada introducida es ilegal, vuelve a solicitar una jugada legal
}
\examples{
tablero <- reiniciar_tablero()
tablero <- turno_humano(tablero, jugada = 4)
visualizar_tablero(tablero)
# tambien se puede hacer lo siguiente (no ejecutar):
# tablero <- reiniciar_tablero()
# tablero <- turno_humano(tablero)
# visualizar_tablero(tablero)
}
