% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generarPosicion.R
\name{crear_posicion_aleatoria}
\alias{crear_posicion_aleatoria}
\title{Crear posición aleatoria}
\usage{
crear_posicion_aleatoria(profundidad = 10)
}
\arguments{
\item{profundidad}{número de jugadas que se simulan aleatoriamente}
}
\value{
Una lista con los siguientes elementos
\itemize{
  \item tablero - matriz 6 x 7 que representa el tablero generado
  \item turnoUltimo - turno del jugador que ha realizado la última jugada en el tablero generado
}
}
\description{
crea un tablero aleatorio de cierta profundidad
}
\examples{
tablero <- crear_posicion_aleatoria(21)
visualizar_tablero(tablero)
sum(tablero == 1)
sum(tablero == 2)
}
