% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profundidadAdaptativa.R
\name{.adaptativa}
\alias{.adaptativa}
\title{profundidad adaptativa}
\usage{
.adaptativa(numJugadas, profundidad, a, p)
}
\description{
profundidad adaptativa
}
\examples{
tablero <- crear_posicion_aleatoria(10)
numJugadas <- seq(7, 1, by = -1/6)
profundidad <- 7
profs <- profundidadAdaptativa(numJugadas, profundidad, a = 0.2, p = 0.8)
(numNodosAdap <- numJugadas^profs)
(numNodosOrig <- numJugadas^profundidad)
ylim = c(0, 1.1 * max(c(numNodosAdap, numNodosOrig)))
plot(numJugadas, profs, type = "b")
plot(numJugadas, numNodosAdap, type = 'b', pch = 19, ylim = ylim)
lines(numJugadas, numNodosOrig, type = 'b', col = "red", pch = 19)
}
