% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluarPosicion.R
\name{evaluar_posicion}
\alias{evaluar_posicion}
\title{evaluación estática de una posición representada en tablero}
\usage{
evaluar_posicion(tablero, turno)
}
\arguments{
\item{turno}{qué jugador comienza pa partida: 1 para humano, 2 para IA.}

\item{profundidad}{profundidad de búsqueda del algoritmo minimax.}
}
\description{
evalua una posición mediante criterios estáticos basados en el
número de casillas conectadas del jugador en turno
}
\examples{
lo siguiente inicia una partida en la que  el jugador humano es el primero en jugar
iniciar_partida(profundidad = 7, turno = 2)

para que sea la IA quien realice la primera jugada:
iniciar_partida(turno = 2)

}
