% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluarPosicion.R
\name{generar_coordenadas_lineas}
\alias{generar_coordenadas_lineas}
\title{evaluación estática de una posición representada en tablero}
\usage{
generar_coordenadas_lineas()
}
\arguments{
\item{tablero}{matriz que representa el estado de un tablero}

\item{turno}{qué jugador comienza pa partida: 1 para humano, 2 para IA.}
}
\description{
evalua una posición mediante criterios estáticos basados en el
número de casillas conectadas del jugador en turno
}
\examples{
tablero <- crear_posicion_aleatoria(5)
visualizar_tablero(tablero)
evaluar_posicion(tablero)
.evaluar_turno(tablero, 1)
.evaluar_turno(tablero, 2)
}
