% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juegoTerminado.R
\name{juego_terminado}
\alias{juego_terminado}
\title{evaluua fin del juego y resultado}
\usage{
juego_terminado(tablero)
}
\arguments{
\item{tablero}{una matriz representado el estado del tablero de juego}
}
\value{
devuelve una lista con el siguiente contenido
\item finalizado: booleano que representa el estado del juego: TRUE finalizado
\item resultado: resultado del juego
}
\description{
evalua una posición mediante criterios estáticos basados en el
número de casillas conectadas del jugador en turno
}
\details{
si finalizado == TRUE, resultado contendrá "GANA HUMANO", "GANA IA" o "TABLAS"·
si finalizado = FALSE, resultado es NA
}
\examples{
tablero <- crear_posicion_aleatoria(21)
juego_terminado(tablero$tablero)
}
