% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juegoTerminado.R
\name{juego_terminado}
\alias{juego_terminado}
\title{end-of-game evaluation and result}
\usage{
juego_terminado(tablero)
}
\arguments{
\item{tablero}{a matrix representing the state of the game board}
}
\value{
returns a list with the following contents
\itemize{
\item{finalizado}: boolean representing the state of the game: TRUE game finished
\item{resultado}: outcome of the game. If the game is over, there are three options
}
}
\description{
evaluates whether the game has ended and with what result 
("HUMAN WINS", "AI WINS", "DRAW").
}
\details{
\itemize{
\item{If the game is over is TRUE}: result can be "WIN HUMAN", "WIN IA" or "DRAW’.
\item{If the game is over is TRUE}: result is NA
}
}
\examples{
tablero <- crear_posicion_aleatoria(21)
p <- visualizar_tablero(tablero)
print(p)
juego_terminado(tablero)
}
