% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jugadas_disponibles}
\alias{jugadas_disponibles}
\title{Columnas disponibles (C++)}
\usage{
jugadas_disponibles(tablero)
}
\arguments{
\item{tablero}{IntegerMatrix de 6 x 7. Estado del tablero.}
}
\value{
IntegerVector con los índices de columna disponibles (1-indexados).
}
\description{
Devuelve los
  índices (1-7) de las columnas que no están completamente llenas.
}
\examples{
tablero <- crear_posicion_aleatoria(20)
jugadas_disponibles(tablero)

}
\seealso{
\code{\link{ordenar_jugadas}}
}
