% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jugadasDisponibles.R
\name{jugadas_disponibles}
\alias{jugadas_disponibles}
\title{Jugadas disponibles}
\usage{
jugadas_disponibles(tablero)
}
\arguments{
\item{tablero}{matriz 6 x 7 que representa la situación del tablero de juego.}
}
\description{
dada una situacvión en el tablero de juego, devuelve las posibles
  jugadas existentes: columnas no completadas
}
\examples{
tablero <- crear_posicion_aleatoria(11)
tablero <- readRDS("tableroPruebas.rds")
visualizar_tablero(tablero)
(jugadas_candidatas <- jugadas_disponibles(tablero))
ordenar_jugadas(tablero, turno = 2)
ordenar_jugadas_cpp(tablero, turno = 2, profundidad = 0)
}
