% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizarTablero.R
\name{visualizar_tablero}
\alias{visualizar_tablero}
\title{visualiza gráficamente la posición de un tablero}
\usage{
visualizar_tablero(tablero)
}
\arguments{
\item{tablero}{matriz 6x7 que representa la posición de un tablero}
}
\value{
devuelve un objeto ggplot listo para representar gráficamente el tablero
}
\description{
representa gráficamente el estado de un "tablero"
}
\examples{
tablero <- crear_posicion_aleatoria(21)
p <- visualizar_tablero(tablero)
print(p)
}
